feat: MVVM 적용 및 디자인 수정
